+++ api/polls/services.py
@@ -1,5 +1,5 @@
 from sqlalchemy import text
-from sqlalchemy.orm import Session
+from sqlalchemy.orm import Session, Query
 from slugify import slugify
 
 from api.polls.schemas import CreatePollSchema, PatchUpdatePollSchema
@@ -9,10 +9,19 @@ from api.users.models import User
 from paginate_sqlalchemy import SqlalchemyOrmPage
 
 
-def get_list_of_polls(db: Session, path: str, page_size, page):
+def get_list_of_all_polls(db: Session, path: str, page_size: int, page: int):
+    query = db.query(Poll).order_by(text('-id'))
+    return get_list_of_polls(path, page_size, page, query)
+
+
+def get_list_of_my_polls(user: User, db: Session, path: str, page_size: int, page: int):
+    query = db.query(Poll).filter_by(creator=user).order_by(text('-id'))
+    return get_list_of_polls(path, page_size, page, query)
+
+
+def get_list_of_polls(path: str, page_size: int, page: int, query: Query):
     """Returns paginated list of polls
     """
-    query = db.query(Poll).order_by(text('-id'))
     path = f"{path}?page_size={page_size}&page="
     page = SqlalchemyOrmPage(query, page=page, items_per_page=page_size)
     next_page = page.next_page
@@ -20,7 +29,7 @@ def get_list_of_polls(db: Session, path: str, page_size, page):
     return {
         'next_page': path + str(next_page) if next_page else None,
         'previous_page': path + str(previous_page) if previous_page else None,
-        'result': page.items,
+        'results': page.items,
         'count': page.item_count
     }